WalletVars and MutableStateCache externalized
